*8759* Move tasks grid to site header

+++ controllers/grid/notifications/NotificationsGridHandler.inc.php
@@ -17,38 +17,19 @@
 import('lib.pkp.classes.controllers.grid.GridHandler');
 
 // Other classes associated with this grid
-import('lib.pkp.controllers.grid.notifications.NotificationsGridRow');
 import('lib.pkp.controllers.grid.notifications.NotificationsGridCellProvider');
 
 class NotificationsGridHandler extends GridHandler {
+	/** @var $_selectedNotificationIds array Set of selected IDs */
+	var $_selectedNotificationIds;
+
 	/**
 	 * Constructor
 	 */
 	function NotificationsGridHandler() {
 		parent::GridHandler();
-
-		$this->addRoleAssignment(
-			array(ROLE_ID_MANAGER, ROLE_ID_SUB_EDITOR, ROLE_ID_ASSISTANT, ROLE_ID_AUTHOR, ROLE_ID_REVIEWER),
-			array('fetchGrid', 'deleteNotification')
-		);
 	}
 
-
-	//
-	// Getters and Setters
-	//
-
-	//
-	// Implement template methods from PKPHandler
-	//
-	/**
-	 * @copydoc PKPHandler::authorize()
-	 */
-	function authorize($request, &$args, $roleAssignments) {
-		import('lib.pkp.classes.security.authorization.PKPSiteAccessPolicy');
-		$this->addPolicy(new PKPSiteAccessPolicy($request, null, $roleAssignments));
-		return parent::authorize($request, $args, $roleAssignments);
-	}
 
 	/**
 	 * @see PKPHandler::initialize()
@@ -56,6 +37,8 @@ class NotificationsGridHandler extends GridHandler {
 	function initialize($request, $args = null) {
 		parent::initialize($request);
 		AppLocale::requireComponents(LOCALE_COMPONENT_APP_SUBMISSION, LOCALE_COMPONENT_PKP_SUBMISSION);
+
+		$this->_selectedNotificationIds = (array) $request->getUserVar('selectedNotificationIds');
 
 		$cellProvider = new NotificationsGridCellProvider();
 		$this->addColumn(
@@ -69,26 +52,97 @@ class NotificationsGridHandler extends GridHandler {
 						'alignment' => COLUMN_ALIGNMENT_LEFT)
 			)
 		);
-		$this->addColumn(
-			new GridColumn(
-				'title',
-				'submission.title',
-				null,
-				'controllers/grid/gridCell.tpl',
-				$cellProvider,
-				array('html' => true,
-						'alignment' => COLUMN_ALIGNMENT_LEFT)
-			)
-		);
 
 		// Set the no items row text
 		$this->setEmptyRowText('grid.noItems');
+
+		import('lib.pkp.classes.linkAction.request.NullAction');
+		$this->addAction(
+			new LinkAction(
+				'markNew',
+				new NullAction(),
+				__('grid.action.markNew'),
+				'edit' // FIXME: Icon
+			),
+			GRID_ACTION_POSITION_BELOW
+		);
+		$this->addAction(
+			new LinkAction(
+				'markRead',
+				new NullAction(),
+				__('grid.action.markRead'),
+				'edit' // FIXME: Icon
+			),
+			GRID_ACTION_POSITION_BELOW
+		);
+
+		$router = $request->getRouter();
+		$this->addAction(
+			new LinkAction(
+				'deleteNotifications',
+				new NullAction(),
+				__('grid.action.delete'),
+				'delete'
+			),
+			GRID_ACTION_POSITION_BELOW
+		);
 	}
 
 
 	//
 	// Overridden methods from GridHandler
 	//
+	/**
+	 * @see GridHandler::getJSHandler()
+	 */
+	public function getJSHandler() {
+		return '$.pkp.controllers.grid.notifications.NotificationsGridHandler';
+	}
+
+	/**
+	 * @see GridHandler::setUrls()
+	 */
+	function setUrls($request) {
+		$router = $request->getRouter();
+		parent::setUrls($request, array(
+			'markNewUrl' => $router->url($request, null, null, 'markNew', null, $this->getRequestArgs()),
+			'markReadUrl' => $router->url($request, null, null, 'markRead', null, $this->getRequestArgs()),
+			'deleteUrl' => $router->url($request, null, null, 'deleteNotifications', null, $this->getRequestArgs()),
+		));
+	}
+
+	/**
+	 * Get the list of "publish data changed" events.
+	 * Used to update the site context switcher upon create/delete.
+	 * @return array
+	 */
+	function getPublishChangeEvents() {
+		return array('updateUnreadNotificationsCount');
+	}
+
+	/**
+	 * @copydoc GridHandler::initFeatures()
+	 */
+	function initFeatures($request, $args) {
+		import('lib.pkp.classes.controllers.grid.feature.selectableItems.SelectableItemsFeature');
+		import('lib.pkp.classes.controllers.grid.feature.PagingFeature');
+		return array(new SelectableItemsFeature(), new PagingFeature());
+	}
+
+	/**
+	 * @copydoc GridHandler::getSelectName()
+	 */
+	function getSelectName() {
+		return 'selectedNotifications';
+	}
+
+	/**
+	 * @copydoc GridHandler::isDataElementSelected()
+	 */
+	function isDataElementSelected($gridDataElement) {
+		return in_array($gridDataElement->getId(), $this->_selectedNotificationIds);
+	}
+
 	/**
 	 * @see GridHandler::loadData()
 	 * @return array Grid data.
@@ -102,7 +156,7 @@ class NotificationsGridHandler extends GridHandler {
 		$rowData = $notifications->toAssociativeArray();
 
 		// Remove not listable task types.
-		$notListableTaskTypes = $this->_getNotListableTaskTypes();
+		$notListableTaskTypes = $this->getNotListableTaskTypes();
 		foreach ($rowData as $key => $notification) {
 			if (in_array($notification->getType(), $notListableTaskTypes)) {
 				unset($rowData[$key]);
@@ -117,39 +171,81 @@ class NotificationsGridHandler extends GridHandler {
 	// Public methods
 	//
 	/**
-	 * Delete a notification
+	 * Mark notifications unread
 	 * @param $args array
 	 * @param $request PKPRequest
+	 * @return string JSON-encoded response
 	 */
-	function deleteNotification($args, $request) {
+	function markNew($args, $request) {
 		$notificationDao = DAORegistry::getDAO('NotificationDAO');
 		$user = $request->getUser();
-		$notification = $notificationDao->getById(
-			$request->getUserVar('notificationId'),
-			$user->getId()
-		);
-		$notificationDao->deleteObject($notification);
+
+		$selectedElements = (array) $request->getUserVar('selectedElements');
+		foreach ($selectedElements as $notificationId) {
+			if ($notification = $notificationDao->getById($notificationId, $user->getId())) {
+				$notificationDao->setDateRead($notificationId, null);
+			}
+		}
+		return DAO::getDataChangedEvent(null, null, $selectedElements);
+	}
+
+	/**
+	 * Mark notifications unread
+	 * @param $args array
+	 * @param $request PKPRequest
+	 * @return string JSON-encoded response
+	 */
+	function markRead($args, $request) {
+		$notificationDao = DAORegistry::getDAO('NotificationDAO');
+		$user = $request->getUser();
+
+		$selectedElements = (array) $request->getUserVar('selectedElements');
+		foreach ($selectedElements as $notificationId) {
+			if ($notification = $notificationDao->getById($notificationId, $user->getId())) {
+				$notificationDao->setDateRead($notificationId, Core::getCurrentDate());
+			}
+		}
+		return DAO::getDataChangedEvent(null, null, $selectedElements);
+	}
+
+	/**
+	 * Delete notifications
+	 * @param $args array
+	 * @param $request PKPRequest
+	 * @return string JSON-encoded response
+	 */
+	function deleteNotifications($args, $request) {
+		$notificationDao = DAORegistry::getDAO('NotificationDAO');
+		$user = $request->getUser();
+
+		$selectedElements = (array) $request->getUserVar('selectedElements');
+		foreach ($selectedElements as $notificationId) {
+			if ($notification = $notificationDao->getById($notificationId, $user->getId())) {
+				$notificationDao->deleteObject($notification);
+			}
+		}
 		return DAO::getDataChangedEvent();
 	}
 
-
-	//
-	// Private helper methods.
-	//
 	/**
-	 * Get the notification types that we don't want
-	 * to list in this grid.
-	 * @return array
+	 * Get unread notifications count
+	 * @param $args array
+	 * @param $request PKPRequest
+	 * @return string JSON-encoded response
 	 */
-	function _getNotListableTaskTypes() {
-		return array(NOTIFICATION_TYPE_SIGNOFF_COPYEDIT, NOTIFICATION_TYPE_SIGNOFF_PROOF);
+	function getUnreadNotificationsCount($args, $request) {
+		$notificationDao = DAORegistry::getDAO('NotificationDAO');
+		$user = $request->getUser();
+		$json = new JSONMessage(true, $notificationDao->getNotificationCount(false, $user->getId(), null, NOTIFICATION_LEVEL_TASK, $this->getNotListableTaskTypes()));
+		return $json->getString();
 	}
 
 	/**
-	 * @copydoc GridHandler::getRowInstance
+	 * Get the notification types that we don't want to list in this grid.
+	 * @return array
 	 */
-	protected function getRowInstance() {
-		return new NotificationsGridRow();
+	static function getNotListableTaskTypes() {
+		return array(NOTIFICATION_TYPE_SIGNOFF_COPYEDIT, NOTIFICATION_TYPE_SIGNOFF_PROOF);
 	}
 }
